development: useSearch moved to hooks
